Tested Working With Permisions

+++ src/resolvers/Mutation.js
@@ -2,22 +2,32 @@ const bcrypt = require('bcryptjs');
 const jwt = require('jsonwebtoken');
 const { randomBytes } = require('crypto');
 const { promisify } = require('util');
+const { transport, makeANiceEmail } = require('../mail');
+const { hasPermission } = require('.../utils');
 require('dotenv').config();
 
 const Mutation = {
-    createItem: async (parent, args, ctx, info) => {
-      // TODO: Check if they are logged in
+    async createItem(parent, args, ctx, info) {
+      if (!ctx.req.userId) {
+        throw new Error('You must be logged in to do that!');
+      }
 
       const item = await ctx.db.mutation.createItem(
         {
-        data: {
-          ...args,
+        data: {  
+          // This is how to create a relationship between the Item and the User
+          user: {
+            connect: {
+              id: ctx.req.userId,
+            },
+          },
+                    ...args,
         },
     }, 
     info
     );
     
-    console.log(item);
+    //console.log(item);
       
     return item;
     },
@@ -40,9 +50,18 @@ const Mutation = {
     async deleteItem(parent, args, ctx, info) {
       const where = { id: args.id };
       // 1. find the item
-      const item = await ctx.db.query.item({ where }, `{ id title}`);
+      const item = await ctx.db.query.item({ where }, `{ id title user { id }}`);
+      console.log("ALLLOOOOO********************************************")
+      console.log(ctx.req.user.name)
       // 2. Check if they own that item, or have the permissions
-      // TODO
+      const ownsItem = item.user.id === ctx.req.userId;
+    const hasPermissions = ctx.req.user.permissions.some(permission =>
+      ['ADMIN', 'ITEMDELETE'].includes(permission)
+    );
+
+    if (!ownsItem && !hasPermissions) {
+      throw new Error("You don't have permission to do that!");
+    }
       // 3. Delete it!
       return ctx.db.mutation.deleteItem({ where }, info);
     },
@@ -69,8 +88,8 @@ const Mutation = {
         httpOnly: true,
         maxAge: 1000 * 60 * 60 * 24 * 365, // 1 year cookie
       });
-      console.log(process.env.APP_SECRET)
-      console.log("Console Log CTX:  " + ctx.res.cookie)
+      //console.log(process.env.APP_SECRET)
+      //console.log("Console Log CTX:  " + ctx.res.cookie)
       // Finalllllly we return the user to the browser
       return user;
     },
@@ -108,14 +127,24 @@ const Mutation = {
       // 2. Set a reset token and expiry on that user
       const randomBytesPromiseified = promisify(randomBytes);
       const resetToken = (await randomBytesPromiseified(20)).toString('hex');
-      const resetTokenExpiry = Date.now() + 3600000; // 1 hour from now
+      const resetTokenExpiry = (Date.now() + 3600000).toString(); // 1 hour from now
       const res = await ctx.db.mutation.updateUser({
         where: { email: args.email },
         data: { resetToken, resetTokenExpiry },
       });
       console.log(res);
-      return { message: 'Thanks!' };
       // 3. Email them that reset token
+      const mailRes = await transport.sendMail({
+        from: 'stephmen@gmail.com',
+        to: user.email,
+        subject: 'Your Password Reset Token',
+        html: makeANiceEmail(`Your Password Reset Token is here!
+        \n\n
+        <a href="${process.env
+          .FRONTEND_URL}/reset?resetToken=${resetToken}">Click Here to Reset</a>`),
+        });
+        //4.Return Message
+        return { message: 'Thanks!' };
     },
     async resetPassword(parent, args, ctx, info) {
       // 1. check if the passwords match
@@ -127,7 +156,7 @@ const Mutation = {
       const [user] = await ctx.db.query.users({
         where: {
           resetToken: args.resetToken,
-          resetTokenExpiry_gte: Date.now() - 3600000,
+          resetTokenExpiry_gte: (Date.now() - 3600000).toString(),
         },
       });
       if (!user) {
@@ -147,12 +176,43 @@ const Mutation = {
       // 6. Generate JWT
       const token = jwt.sign({ userId: updatedUser.id }, process.env.APP_SECRET);
       // 7. Set the JWT cookie
-      ctx.response.cookie('token', token, {
+      ctx.res.cookie('token', token, {
         httpOnly: true,
         maxAge: 1000 * 60 * 60 * 24 * 365,
       });
       // 8. return the new user
       return updatedUser;
+    },
+    async updatePermissions(parent, args, ctx, info) {
+      // 1. Check if they are logged in
+      if (!ctx.req.userId) {
+        throw new Error('You must be logged in!');
+      }
+      // 2. Query the current user
+      const currentUser = await ctx.db.query.user(
+        {
+          where: {
+            id: ctx.req.userId,
+          },
+        },
+        info
+      );
+      // 3. Check if they have permissions to do this
+      hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE']);
+      // 4. Update the permissions
+      return ctx.db.mutation.updateUser(
+        {
+          data: {
+            permissions: {
+              set: args.permissions,
+            },
+          },
+          where: {
+            id: args.userId,
+          },
+        },
+        info
+      );
     },
   };
